+++ src/main/java/kusitms/backend/user/infra/sms/SmsManager.java
@@ -1,13 +1,13 @@
-package kusitms.backend.user.application;
+package kusitms.backend.user.infra.sms;
 
 import com.twilio.Twilio;
 import com.twilio.rest.api.v2010.account.Message;
 import com.twilio.type.PhoneNumber;
 import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
-@Service
+@Component
 public class SmsManager {
 
     @Value("${twilio.account_sid}")
